Break out the ApplicationInProgress component
