+++ src/applications/personalization/dashboard-2/components/apply-for-benefits/ApplyForBenefits.jsx
@@ -8,50 +8,10 @@ import {
   formTitles,
   isSIPEnabledForm,
   presentableFormIDs,
-  recordDashboardClick,
   sipFormSorter,
 } from '~/applications/personalization/dashboard/helpers';
 
-const ApplicationInProgress = ({
-  continueUrl,
-  expirationDate,
-  formId,
-  formTitle,
-  lastOpenedDate,
-}) => {
-  return (
-    <div className="vads-u-display--flex vads-l-col--12 medium-screen:vads-l-col--6 small-desktop-screen:vads-l-col--4 medium-screen:vads-u-padding-right--3 vads-u-padding-bottom--3">
-      <div className="vads-u-display--flex vads-u-width--full vads-u-flex-direction--column vads-u-justify-content--space-between vads-u-align-items--flex-start vads-u-background-color--gray-lightest vads-u-padding--2p5">
-        <div>
-          <div style={{ textTransform: 'uppercase' }}>{formId}</div>
-          <h4 className="vads-u-font-size--h3 vads-u-margin-top--0">
-            {formTitle}
-          </h4>
-          <div className="vads-u-display--flex">
-            <i
-              aria-hidden="true"
-              className={`fas fa-fw fa-exclamation-circle vads-u-margin-right--1 vads-u-margin-top--0p5`}
-            />
-            <div>
-              <p className="vads-u-margin-top--0">
-                Application expires on: {expirationDate}
-              </p>
-              <p>Last opened on: {lastOpenedDate}</p>
-            </div>
-          </div>
-        </div>
-        <a
-          className="usa-button usa-button-primary"
-          aria-label={`Continue your ${formTitle}`}
-          href={continueUrl}
-          onClick={recordDashboardClick(formId, 'continue-button')}
-        >
-          Continue your application
-        </a>
-      </div>
-    </div>
-  );
-};
+import ApplicationInProgress from './ApplicationInProgress';
 
 const ApplyForBenefits = ({ savedForms }) => {
   return (
